CBAPI-1562 - clear up time zone confusion in date and time inputs
Explicit assumptions have been added to the state manager APIs that allow
date/time input, that said date and time is in the _local_ timezone. SQLite
likes to store time data in UTC, but it's best equipped to handle the
conversion itself (using the 'utc' modifier to its date and time functions).
Modified the SQL it uses to take this into account.